#1363 Move DataSourceMapConverter to transaction package.
