+++ sharding-core/src/test/java/io/shardingsphere/spi/transaction/xa/SPIDataSourceMapConverterTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.spi.xa;
+package io.shardingsphere.spi.transaction.xa;
 
 import io.shardingsphere.core.constant.DatabaseType;
 import org.junit.Test;
