+++ sharding-core/src/test/resources/META-INF/services/io.shardingsphere.spi.transaction.xa.DataSourceMapConverter
@@ -1,0 +1,1 @@
+io.shardingsphere.spi.transaction.xa.FixedDataSourceMapConverter
