--- sharding-core/src/test/resources/META-INF/services/io.shardingsphere.spi.xa.DataSourceMapConverter
@@ -1,1 +1,0 @@
-io.shardingsphere.spi.xa.FixedBackendDataSourceFactory
